Certain args, kwargs of sink functions are affected by taint
The aim of this is to reduce the number of false positives.

In the json trigger file you can specify a list of arg positions or
keywords which propagate or ignore taint.

So e.g there may be an execute function
where the first argument is raw SQL query text which is affected by
taint, but any other arguments are variables passed in to the prepared SQL
statement and (we assume) can be tainted user input without a problem.

To do this we can no longer just use the RHSVisitor to see which
variables are used by a sink. We need the results of each arg / kwarg
separately.

There is the added complication of *args and **kwargs where we need to
determine if a RHS variable could pass into a propagating arg / kwarg or not.
Probably the best way to understand it is to look at the added test cases.