Imports: --dont-prepend-root flag
For a project with root in /app,
currently pyt expects that all of the imports are of the form:

from app.folder.module import thing

But actually a project could not be expecting the module root to
be prepended.

My projects use:

from folder.module import thing

This adds an optional boolean flag to change the behaviour of
get_modules.