Remove extraneous reassignments in output
The output should consist of the path from the source to the sink.

Anything which happens after the source reaches the sink is irrelevant
and just makes the output longer and confusing to interpret.

None of the lines removed from the tests actually affected the
vulnerability chain.

Perhaps this should be dealt with somewhere in the definition_chain or
vulnerability functions: here we just trim the chain upon reaching the
sink in the vulnerability_helper.