Simplify trigger file for sink argument propagation
This changes the schema of the trigger file.

Previously there were too many options and it was confusing. My fault,
sorry.

This meant that `db.execute(query, **TAINT)` was marked as a
vulnerability whereas `db.execute(text=query, **TAINT)` wasn't.

Neither are vulnerabilities, so this gave a FALSE POSITIVE.

Now we have `arg_dict` which is a dictionary of keyword to argument
position.

E.g. for `def f(a, b, *, c)` we can specify the arg_dict as:

```
{"a": 0, "b": 1, "c": null}
```

if we want them all to propagate or not propagate depending on the
`unlisted_args_propagate` value.

This way, we can more easily define db.execute as:

```
            "execute(": {
                "unlisted_args_propagate": false,
                "arg_dict": {
                    "text": 0
                }
            },
```